@@ -1,12 +1,12 @@
 name: dokit
 description: 开发工具集
-version: 0.6.0
+version: 0.8.0
 homepage: https://www.dokit.cn/#/index/home
 
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
