@@ -20,25 +20,25 @@ typedef LogCallback = void Function(String);
 typedef ExceptionCallback = void Function(dynamic, StackTrace);
 
 const String DK_PACKAGE_NAME = 'dokit';
-const String DK_PACKAGE_VERSION = '0.6.0';
+const String DK_PACKAGE_VERSION = '0.8.0';
 
 //默认release模式不开启该功能
 const bool release = kReleaseMode;
 
 //记录当前zone
-Zone _zone;
+Zone? _zone;
 
 // ignore: avoid_classes_with_only_static_members
 class DoKit {
   // 初始化方法,app或者appCreator必须设置一个
   static Future<void> runApp(
-      {DoKitApp app,
-      DoKitAppCreator appCreator,
+      {DoKitApp? app,
+      DoKitAppCreator? appCreator,
       bool useInRelease = false,
-      LogCallback logCallback,
-      ExceptionCallback exceptionCallback,
+      LogCallback? logCallback,
+      ExceptionCallback? exceptionCallback,
       List<String> methodChannelBlackList = const <String>[],
-      Function releaseAction}) async {
+      Function? releaseAction}) async {
     // 统计用户信息，便于了解该开源产品的使用量 (请大家放心，我们不用于任何恶意行为)
     upLoadUserInfo();
 
@@ -51,7 +51,7 @@ class DoKit {
         if (app != null) {
           dart.runApp(app.origin);
         } else {
-          dart.runApp((await appCreator())?.origin);
+          dart.runApp((await appCreator!()).origin);
         }
       }
       return;
@@ -60,7 +60,7 @@ class DoKit {
     await runZoned(
       () async => <void>{
         _ensureDoKitBinding(useInRelease: useInRelease),
-        _runWrapperApp(app != null ? app : await appCreator()),
+        _runWrapperApp(app != null ? app : await appCreator!()),
         _zone = Zone.current
       },
       zoneSpecification: ZoneSpecification(
@@ -92,7 +92,7 @@ void _ensureDoKitBinding({bool useInRelease = false}) {
 void _runWrapperApp(IDoKitApp wrapper) {
   DoKitWidgetsFlutterBinding.ensureInitialized()
 // ignore: invalid_use_of_protected_member
-    ..scheduleAttachRootWidget(wrapper)
+    ?..scheduleAttachRootWidget(wrapper)
     ..scheduleWarmUpFrame();
   addEntrance();
 }
@@ -101,21 +101,21 @@ void _collectLog(String line) {
   LogManager.instance.addLog(LogBean.TYPE_INFO, line);
 }
 
-void _collectError(Object details, Object stack) {
+void _collectError(Object? details, Object? stack) {
   LogManager.instance.addLog(
       LogBean.TYPE_ERROR, '${details?.toString()}\n${stack?.toString()}');
 }
 
 void addEntrance() {
-  WidgetsBinding.instance.addPostFrameCallback((_) {
+  WidgetsBinding.instance?.addPostFrameCallback((_) {
     final DoKitBtn floatBtn = DoKitBtn();
     floatBtn.addToOverlay();
     KitPageManager.instance.loadCache();
   });
 }
 
-void dispose({@required BuildContext context}) {
-  doKitOverlayKey.currentState.widget.initialEntries.forEach((element) {
+void dispose({required BuildContext context}) {
+  doKitOverlayKey.currentState?.widget.initialEntries.forEach((element) {
 // element.remove();
   });
 }
@@ -126,7 +126,7 @@ void upLoadUserInfo() async {
   final request = await client.postUrl(Uri.parse(url));
   final packageInfo = await PackageInfo.fromPlatform();
 
-  Locale locale;
+  Locale? locale;
   void finder(Element element) {
     if (element.widget is Localizations) {
       locale ??= (element.widget as Localizations).locale;
@@ -135,7 +135,7 @@ void upLoadUserInfo() async {
     }
   }
 
-  DoKitApp.appKey.currentContext.visitChildElements(finder);
+  DoKitApp.appKey.currentContext?.visitChildElements(finder);
 
   final appId = packageInfo.packageName;
   // 在iOS上可能获取不到appName
