@@ -7,14 +7,14 @@ import 'package:string_scanner/string_scanner.dart';
 
 class SyntaxHighlighterStyle {
   SyntaxHighlighterStyle(
-      {this.baseStyle,
-        this.numberStyle,
-        this.commentStyle,
-        this.keywordStyle,
-        this.stringStyle,
-        this.punctuationStyle,
-        this.classStyle,
-        this.constantStyle});
+      {required this.baseStyle,
+      required this.numberStyle,
+      required this.commentStyle,
+      required this.keywordStyle,
+      required this.stringStyle,
+      required this.punctuationStyle,
+      required this.classStyle,
+      required this.constantStyle});
 
   static SyntaxHighlighterStyle lightThemeStyle() {
     return SyntaxHighlighterStyle(
@@ -61,7 +61,7 @@ class DartSyntaxHighlighter extends SyntaxHighlighter {
     _style ??= SyntaxHighlighterStyle.darkThemeStyle();
   }
 
-  SyntaxHighlighterStyle _style;
+  SyntaxHighlighterStyle? _style;
 
   static const List<String> _keywords = const <String>[
     'abstract',
@@ -126,10 +126,10 @@ class DartSyntaxHighlighter extends SyntaxHighlighter {
     'bool'
   ];
 
-  String _src;
-  StringScanner _scanner;
+  late String _src;
+  late StringScanner _scanner;
 
-  List<_HighlightSpan> _spans;
+  late List<_HighlightSpan> _spans;
 
   @override
   TextSpan format(String src) {
@@ -142,24 +142,26 @@ class DartSyntaxHighlighter extends SyntaxHighlighter {
       int currentPosition = 0;
 
       for (_HighlightSpan span in _spans) {
-        if (currentPosition != span.start)
+        if (currentPosition != span.start) {
           formattedText
               .add(TextSpan(text: _src.substring(currentPosition, span.start)));
+        }
 
         formattedText.add(TextSpan(
-            style: span.textStyle(_style), text: span.textForSpan(_src)));
+            style: span.textStyle(_style!), text: span.textForSpan(_src)));
 
         currentPosition = span.end;
       }
 
-      if (currentPosition != _src.length)
+      if (currentPosition != _src.length) {
         formattedText
             .add(TextSpan(text: _src.substring(currentPosition, _src.length)));
+      }
 
-      return TextSpan(style: _style.baseStyle, children: formattedText);
+      return TextSpan(style: _style!.baseStyle, children: formattedText);
     } else {
       // Parsing failed, return with only basic formatting
-      return TextSpan(style: _style.baseStyle, text: src);
+      return TextSpan(style: _style!.baseStyle, text: src);
     }
   }
 
@@ -173,18 +175,18 @@ class DartSyntaxHighlighter extends SyntaxHighlighter {
       // Block comments
       if (_scanner.scan(RegExp(r'/\*(.|\n)*\*/'))) {
         _spans.add(_HighlightSpan(_HighlightType.comment,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Line comments
       if (_scanner.scan('//')) {
-        final int startComment = _scanner.lastMatch.start;
+        final int startComment = _scanner.lastMatch!.start;
 
         bool eof = false;
         int endComment;
         if (_scanner.scan(RegExp(r'.*\n'))) {
-          endComment = _scanner.lastMatch.end - 1;
+          endComment = _scanner.lastMatch!.end - 1;
         } else {
           eof = true;
           endComment = _src.length;
@@ -201,94 +203,95 @@ class DartSyntaxHighlighter extends SyntaxHighlighter {
       // Raw r"String"
       if (_scanner.scan(RegExp(r'r".*"'))) {
         _spans.add(_HighlightSpan(_HighlightType.string,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Raw r'String'
       if (_scanner.scan(RegExp(r"r'.*'"))) {
         _spans.add(_HighlightSpan(_HighlightType.string,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Multiline """String"""
       if (_scanner.scan(RegExp(r'"""(?:[^"\\]|\\(.|\n))*"""'))) {
         _spans.add(_HighlightSpan(_HighlightType.string,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Multiline '''String'''
       if (_scanner.scan(RegExp(r"'''(?:[^'\\]|\\(.|\n))*'''"))) {
         _spans.add(_HighlightSpan(_HighlightType.string,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // "String"
       if (_scanner.scan(RegExp(r'"(?:[^"\\]|\\.)*"'))) {
         _spans.add(_HighlightSpan(_HighlightType.string,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // 'String'
       if (_scanner.scan(RegExp(r"'(?:[^'\\]|\\.)*'"))) {
         _spans.add(_HighlightSpan(_HighlightType.string,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Double
       if (_scanner.scan(RegExp(r'\d+\.\d+'))) {
         _spans.add(_HighlightSpan(_HighlightType.number,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Integer
       if (_scanner.scan(RegExp(r'\d+'))) {
         _spans.add(_HighlightSpan(_HighlightType.number,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Punctuation
       if (_scanner.scan(RegExp(r'[\[\]{}().!=<>&\|\?\+\-\*/%\^~;:,]'))) {
         _spans.add(_HighlightSpan(_HighlightType.punctuation,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Meta data
       if (_scanner.scan(RegExp(r'@\w+'))) {
         _spans.add(_HighlightSpan(_HighlightType.keyword,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+            _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         continue;
       }
 
       // Words
       if (_scanner.scan(RegExp(r'\w+'))) {
-        _HighlightType type;
+        _HighlightType? type;
 
-        String word = _scanner.lastMatch[0];
+        String word = _scanner.lastMatch![0]!;
         if (word.startsWith('_')) word = word.substring(1);
 
-        if (_keywords.contains(word))
+        if (_keywords.contains(word)) {
           type = _HighlightType.keyword;
-        else if (_builtInTypes.contains(word))
+        } else if (_builtInTypes.contains(word)) {
           type = _HighlightType.keyword;
-        else if (_firstLetterIsUpperCase(word))
+        } else if (_firstLetterIsUpperCase(word)) {
           type = _HighlightType.klass;
-        else if (word.length >= 2 &&
+        } else if (word.length >= 2 &&
             word.startsWith('k') &&
-            _firstLetterIsUpperCase(word.substring(1)))
+            _firstLetterIsUpperCase(word.substring(1))) {
           type = _HighlightType.constant;
+        }
 
         if (type != null) {
           _spans.add(_HighlightSpan(
-              type, _scanner.lastMatch.start, _scanner.lastMatch.end));
+              type, _scanner.lastMatch!.start, _scanner.lastMatch!.end));
         }
       }
 
@@ -345,21 +348,22 @@ class _HighlightSpan {
   }
 
   TextStyle textStyle(SyntaxHighlighterStyle style) {
-    if (type == _HighlightType.number)
+    if (type == _HighlightType.number) {
       return style.numberStyle;
-    else if (type == _HighlightType.comment)
+    } else if (type == _HighlightType.comment) {
       return style.commentStyle;
-    else if (type == _HighlightType.keyword)
+    } else if (type == _HighlightType.keyword) {
       return style.keywordStyle;
-    else if (type == _HighlightType.string)
+    } else if (type == _HighlightType.string) {
       return style.stringStyle;
-    else if (type == _HighlightType.punctuation)
+    } else if (type == _HighlightType.punctuation) {
       return style.punctuationStyle;
-    else if (type == _HighlightType.klass)
+    } else if (type == _HighlightType.klass) {
       return style.classStyle;
-    else if (type == _HighlightType.constant)
+    } else if (type == _HighlightType.constant) {
       return style.constantStyle;
-    else
+    } else {
       return style.baseStyle;
+    }
   }
 }
